Update advanced example
This advanced example includes using the addon ofxWinDialog.
https://github.com/leadedge/ofxWinDialog
Two dialogs are created :
  - example dialog with controls
  - About dialog with Version information
Resources are used to embed data into the executable instead of
requiring additional files at runtime.
  - Image
  - Icon
The resource also contains Version information which enables
the version to show in Windows explorer and in the file properties.
See resource.h and resource.rc.
Menu and Dialog Settings are saved and restored with
an initialization file using Save and Load functions.